fixed tests

+++ src/main/scala/is/hail/expr/types/TFloat64.scala
@@ -25,7 +25,7 @@ class TFloat64(override val required: Boolean) extends TNumeric {
   override def valuesSimilar(a1: Annotation, a2: Annotation, tolerance: Double): Boolean =
     a1 == a2 || (a1 != null && a2 != null &&
       (math.abs(a1.asInstanceOf[Double] - a2.asInstanceOf[Double]) <= tolerance ||
-        D_==(a1.asInstanceOf[Float], a2.asInstanceOf[Float], tolerance) ||
+        D_==(a1.asInstanceOf[Double], a2.asInstanceOf[Double], tolerance) ||
         (a1.asInstanceOf[Double].isNaN && a2.asInstanceOf[Double].isNaN)))
 
   override def scalaClassTag: ClassTag[java.lang.Double] = classTag[java.lang.Double]
